fix (#203)
* fix

* fix

* fix order response

+++ LgymApi.Application/Training/Models/ScoreResult.cs
@@ -4,7 +4,7 @@ namespace LgymApi.Application.Features.Training.Models;
 
 public sealed class ScoreResult
 {
-    public int Reps { get; init; }
+    public double Reps { get; init; }
     public double Weight { get; init; }
     public WeightUnits Unit { get; init; }
 }
